Добавит форматирование продолжитеьности в компонентах
