+++ src/utils/const.js
@@ -7,3 +7,16 @@ export const EvtKey = {
 };
 
 export const EMOTIONS = ['smile', 'sleeping', 'puke', 'angry'];
+
+export const StringFormats = {
+  RELEASE_DATE: 'DD MMMM YYYY',
+  COMMENT_DATE: 'YYYY/MM/D HH:mm',
+  RUNTIME_MINUTES: 'mm[M]',
+  RUNTIME_HOURS: 'H[h] mm[m]',
+};
+
+export const Runtime = {
+  MIN: 50,
+  MAX: 180,
+  MINUTES_IN_HOUR: 60,
+};
